Allow running Spock without Qt bindings
